% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmodes_gower.R
\name{predict.kmodes_gower}
\alias{predict.kmodes_gower}
\title{Predict KModes Gower Object on New Data}
\usage{
\method{predict}{kmodes_gower}(object, newdata, type = c("cluster",
  "distance"), normalize = TRUE)
}
\arguments{
\item{object}{\code{kmodes_gower} object produces by \link{kmodes_gower} function.}

\item{newdata}{data.frame on which to predict.}

\item{type}{What to predict on the new data. Either the assigned cluster or the distance matrix
containining the distance for each objservation to each cluster mode.}

\item{normalize}{a logical value for whether \code{rowSums(dists)} should all equal one.}
}
\description{
Predict KModes Gower Object on New Data
}
