% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_kmodes.R
\name{predict.sparsekmodes}
\alias{predict.sparsekmodes}
\title{Predict Sparse K-Modes Object on New Data}
\usage{
\method{predict}{sparsekmodes}(object, newdata, type = c("cluster", "sim"),
  normalize = TRUE)
}
\arguments{
\item{object}{an object of class "sparsekmodes"}

\item{newdata}{A sparse, binary matrix with the same number of columns as \code{nrow(object@centers)}}

\item{type}{what type of prediction to calculate. Cluster returns the id of the closest center.
sim returns the similarity matrix between \code{newdata} and \code{object@centers}.}

\item{normalize}{Whether to normalize the similarity matrix such that \code{all(rowSums(sim) == 1)} is TRUE.}
}
\description{
Predict Sparse K-Modes Object on New Data
}
